ubports-qa: run a full upgrade on install
The `apt upgrade` command never removes installed packages, meaning that
if a package from the PPA conflicts with installed packages, it won't be
installed or upgraded.

The `apt full-upgrade` command, on the other hand, allows installing or
upgrading all the packages from the PPA, even if that requires removing
some of the installed packages.